+++ release-notes.md
@@ -9,6 +9,7 @@ Version 1.4.1 (unreleased)
 * Added GridCell.getId() that maps to Grid.cellToId()
 * Added Vec3d.toVec3i(), floor(), and ceil() for Vec3d to Vec3i conversion.
 * Added Grid.getContainingCell(Vec3i)
+* Added Vec3d.distance(x,y,z) and Vec3d.distanceSq(x,y,z)
 
 
 Version 1.4.0 (latest)